Finished testing and updated pycodestyle

+++ app.py
@@ -1,5 +1,6 @@
 import os
-from flask import Flask, jsonify, redirect, render_template, session, url_for, request, abort
+from flask import (Flask, jsonify, redirect, render_template, session,
+                   url_for, request, abort)
 import json
 from models import setup_db
 from flask_cors import CORS
@@ -8,11 +9,13 @@ from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
 from models import db, Tune, Composer, Mastery, Key, Playlist, Playlist_Tune
 
+
 def create_app():
     app = Flask(__name__)
     setup_db(app)
     CORS(app)
     return app
+
 
 app = create_app()
 migrate = Migrate(app, db)
@@ -25,10 +28,11 @@ def return_composer(composer_name):
         try:
             new_composer.insert()
             composer_entry = Composer.query.filter_by(name=composer_name)\
-                                        .first()
+                .first()
         except Exception as e:
-            abort(400, "Composer failed to insert!")
+            abort(400, 'Composer failed to insert!')
     return composer_entry
+
 
 def tunes_to_dict(tune_list):
     library = {}
@@ -46,7 +50,7 @@ def index():
 
 
 @app.route('/tunes/', methods=['GET'])
-@requires_auth("get:tunes")
+@requires_auth('get:tunes')
 def all_tunes(jwt):
     # Returns a list of all tunes from database.
     try:
@@ -65,14 +69,14 @@ def all_tunes(jwt):
 
 
 @app.route('/tunes/<id>/', methods=['GET'])
-@requires_auth("get:tunes")
+@requires_auth('get:tunes')
 def single_tune_info(jwt, id):
     try:
         id = int(id)
         tune = Tune.query.filter_by(id=id).first()
 
         if tune is None:
-            abort(404, "Tune with this id not found!")
+            abort(404, 'Tune with this id not found!')
 
         library = tunes_to_dict([tune])
         return json.dumps({
@@ -80,18 +84,18 @@ def single_tune_info(jwt, id):
             "tune": library
         }), 200
     except ValueError:
-        abort(400, "Tune id must be an integer")
+        abort(400, 'Tune id must be an integer')
     except Exception as e:
         if tune is None:
             print(e)
             abort(404)
-        else: 
+        else:
             print(e)
             abort(400)
 
 
 @app.route('/tunes/', methods=['POST'])
-@requires_auth("post:tunes")
+@requires_auth('post:tunes')
 def add_tune(jwt):
     # Retrieves the tune metadata from the request.
     tune_data = json.loads(request.data.decode('utf-8'))
@@ -99,12 +103,12 @@ def add_tune(jwt):
     # Determines whether or not the tune is already in the database.
     tune_exists = Tune.query.filter_by(title=tune_data['title']).first()
     if tune_exists is not None:
-        abort(409, "A tune with this title already exists!")
+        abort(409, 'A tune with this title already exists!')
 
     # Retrieves the composer and key objects, for determining their IDs.
     composer = return_composer(tune_data['composer'])
     key = Key.query.filter_by(key=tune_data['key']).first()
-    
+
     # Builds the new tune object and insert into the database.
     new_tune = Tune(title=tune_data['title'],
                     composer=composer.id,
@@ -114,14 +118,14 @@ def add_tune(jwt):
     try:
         new_tune.insert()
         return json.dumps({'success': True,
-                        'new tune': new_tune.format()})
+                           'new tune': new_tune.format()})
     except Exception as e:
         print(e)
         abort(400)
 
 
 @app.route('/tunes/<id>', methods=['PATCH'])
-@requires_auth("patch:tunes")
+@requires_auth('patch:tunes')
 def edit_tune(jwt, id):
     data = json.loads(request.data.decode('utf-8'))
     tune = Tune.query.filter_by(id=id).first()
@@ -130,24 +134,29 @@ def edit_tune(jwt, id):
     if data.get('title'):
         tune.title = data['title']
     if data.get('composer'):
-        tune.composer = data['composer']
+        composer = return_composer(data['composer'])
+        tune.composer = composer.id
     if data.get('key'):
-        tune.key = data['key']
+        key = Key.query.filter_by(key=data['key']).first()
+        tune.key = key.id
     if data.get('mastery'):
         tune.mastery = data['mastery']
     try:
         tune.update()
-        return json.dumps({'success': True, "updated tune": tune.format()}), 200
+        return json.dumps({'success': True,
+                           'updated tune': tune.format()}), 200
     except Exception as e:
         print(e)
         abort(400)
 
 
 @app.route('/tunes/<id>', methods=['DELETE'])
-@requires_auth("delete:tunes")
+@requires_auth('delete:tunes')
 def delete_tune(jwt, id):
     tune = Tune.query.filter_by(id=id).first()
 
+    if tune is None:
+        abort(422, 'The tune doesn't exist!')
     try:
         tune.delete()
         db.session.commit()
@@ -157,31 +166,13 @@ def delete_tune(jwt, id):
     return json.dumps({'success': True, "deleted tune": tune.format()})
 
 
-@app.errorhandler(422)
-def unprocessable(error):
-    return jsonify({
-                    "success": False,
-                    "error": 422,
-                    "message": "unprocessable"
-                    }), 422
-
-
-@app.errorhandler(404)
-def resource_not_found(error):
-    return jsonify({
-                    "success": False,
-                    "error": 404,
-                    "message": "resource not found"
-                    }), 404
-
-
 @app.errorhandler(400)
 def resource_not_found(error):
     return jsonify({
                     "success": False,
                     "error": 400,
                     "message": "bad request"
-                    }), 400
+    }), 400
 
 
 @app.errorhandler(401)
@@ -193,6 +184,33 @@ def auth_error_handler(error):
     }), 401
 
 
+@app.errorhandler(403)
+def auth_error_handler(error):
+    return jsonify({
+                    "success": False,
+                    "error": 403,
+                    "message": "forbidden"
+    }), 403
+
+
+@app.errorhandler(404)
+def resource_not_found(error):
+    return jsonify({
+                    "success": False,
+                    "error": 404,
+                    "message": "resource not found"
+    }), 404
+
+
+@app.errorhandler(405)
+def resource_not_found(error):
+    return jsonify({
+                    "success": False,
+                    "error": 405,
+                    "message": "method not allowed"
+    }), 405
+
+
 @app.errorhandler(409)
 def auth_error_handler(error):
     return jsonify({
@@ -200,6 +218,15 @@ def auth_error_handler(error):
                     "error": 409,
                     "message": "conflict"
     }), 409
+
+
+@app.errorhandler(422)
+def unprocessable(error):
+    return jsonify({
+                    "success": False,
+                    "error": 422,
+                    "message": "unprocessable"
+    }), 422
 
 
 if __name__ == '__main__':
